Minor fixes

+++ common/Sparse_to_Dense.py
@@ -33,7 +33,7 @@ def to_dense(wavelength_list, ratios, relative_humidity, res_array):
     zeros_arr = np.zeros((len(wavelength_list), ratios.size, ratios.size, ratios.size, ratios.size, ratios.size,
                           ratios.size, ratios.size,
                           len(relative_humidity), res_array['thetas'].size, res_array['tau'].size,
-                          res_array['rho_s'].size))
+                          res_array['rho_s'].size), dtype="float32")
     data = xr.DataArray(zeros_arr,
                         dims=(
                         "wavelength", "rDU", "rBC", "rOM", "rSS", "rSU", "rNI", "rAM", "RH", "thetas", "tau", "rho_s"),
@@ -107,6 +107,7 @@ if test is not None:
     wavelength_list = [0.56]
     relative_humidity = [0.3]
     ratios = Ratio.Ratio(1)
+
     data = to_dense(wavelength_list,ratios,relative_humidity,demo)
 
     print(data.sel(wavelength=0.56, rDU=0, rBC=0, rOM=0, rSS=0, rSU=0, rNI=0, rAM=1, RH=0.3, thetas=60, tau=0.8, rho_s=0.7).values)
